refactor confirmation modal into pieces before creating the error content
