+++ src/components/TransactionConfirmationModal/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../../tsconfig.strict.json",
+  "include": ["**/*"]
+}
